+++ lib/checkin_handler.py
@@ -7,12 +7,15 @@ from typing import TYPE_CHECKING
 
 from .flight import Flight
 from .general import CheckInError, make_request
+from .log import get_logger
 
 if TYPE_CHECKING:  # pragma: no cover
     from .checkin_scheduler import CheckInScheduler
 
 CHECKIN_URL = "mobile-air-operations/v1/mobile-air-operations/page/check-in/"
 MANUAL_CHECKIN_URL = "https://mobile.southwest.com/check-in"
+
+logger = get_logger(__name__)
 
 
 class CheckInHandler:
@@ -31,6 +34,7 @@ class CheckInHandler:
         self.last_name = self.checkin_scheduler.flight_retriever.last_name
 
     def schedule_check_in(self) -> None:
+        logger.debug("Scheduling check-in for current flight")
         process = Process(target=self._set_check_in)
         process.start()
 
@@ -43,6 +47,7 @@ class CheckInHandler:
     def _wait_for_check_in(self, checkin_time: datetime) -> None:
         current_time = datetime.utcnow()
         if checkin_time <= current_time:
+            logger.debug("Check-in time has passed. Going straight to check-in")
             return
 
         # Refresh headers 10 minutes before to make sure they are valid
@@ -50,12 +55,27 @@ class CheckInHandler:
 
         # Only try to refresh the headers if the checkin is more than ten minutes away
         if sleep_time > 0:
-            time.sleep(sleep_time)
+            logger.debug("Sleeping until ten minutes before check-in...")
+            self.safe_sleep(sleep_time)
             self.checkin_scheduler.refresh_headers()
 
         current_time = datetime.utcnow()
         sleep_time = (checkin_time - current_time).total_seconds()
+        logger.debug("Sleeping until check-in: %d seconds...", sleep_time)
         time.sleep(sleep_time)
+
+    @staticmethod
+    def safe_sleep(total_sleep_time: int) -> None:
+        """
+        If the total sleep time is too long, an overflow error could occur.
+        Therefore, the script will continuously sleep in two week periods
+        to avoid this issue.
+        """
+        two_weeks = 60 * 60 * 24 * 14
+        while total_sleep_time > 0:
+            sleep_time = min(total_sleep_time, two_weeks)
+            time.sleep(sleep_time)
+            total_sleep_time -= sleep_time
 
     def _check_in(self) -> None:
         """
@@ -63,10 +83,11 @@ class CheckInHandler:
         a POST request to submit the check in.
         """
         account_name = f"{self.first_name} {self.last_name}"
+        logger.debug("Attempting to check in")
         print(
-            f"Checking in to flight from '{self.flight.departure_airport}' to '{self.flight.destination_airport}' "
-            f"for {account_name}\n"
-        )
+            f"Checking in to flight from '{self.flight.departure_airport}' to "
+            f"'{self.flight.destination_airport}' for {account_name}\n"
+        )  # Don't log as it has sensitve information
 
         headers = self.checkin_scheduler.headers
         info = {
@@ -76,16 +97,20 @@ class CheckInHandler:
         site = CHECKIN_URL + self.flight.confirmation_number
 
         try:
+            logger.debug("Making GET request to check in")
             response = make_request("GET", site, headers, info)
 
             info = response["checkInViewReservationPage"]["_links"]["checkIn"]
             site = f"mobile-air-operations{info['href']}"
 
+            logger.debug("Making POST request to check in")
             reservation = make_request("POST", site, headers, info["body"])
         except CheckInError as err:
+            logger.debug("Failed to check in. Error: %s. Exiting", err)
             self.notification_handler.failed_checkin(err, self.flight)
             return
 
+        logger.debug("Successfully checked in!")
         self.notification_handler.successful_checkin(
             reservation["checkInConfirmationPage"], self.flight
         )