+++ lib/notification_handler.py
@@ -6,11 +6,13 @@ import apprise
 
 from .flight import Flight
 from .general import CheckInError, LoginError, NotificationLevel
+from .log import get_logger
 
 if TYPE_CHECKING:  # pragma: no cover
     from .flight_retriever import FlightRetriever
 
 MANUAL_CHECKIN_URL = "https://mobile.southwest.com/check-in"
+logger = get_logger(__name__)
 
 
 class NotificationHandler:
@@ -25,7 +27,7 @@ class NotificationHandler:
         return f"{self.flight_retriever.first_name} {self.flight_retriever.last_name}"
 
     def send_notification(self, body: str, level: int = None) -> None:
-        print(body)
+        print(body)  # This isn't logged as it contains sensitive information
 
         # Check the level to see if we still want to send it. If level is none, it means
         # the message will always be printed. For example, this is used when testing notifications.
@@ -42,10 +44,17 @@ class NotificationHandler:
         if len(flights) == 0:
             return
 
-        flight_schedule_message = f"Successfully scheduled the following flights to check in for {self._get_account_name()}:\n"
+        flight_schedule_message = (
+            f"Successfully scheduled the following flights to check in for "
+            f"{self._get_account_name()}:\n"
+        )
         for flight in flights:
-            flight_schedule_message += f"Flight from {flight.departure_airport} to {flight.destination_airport} at {flight.departure_time} UTC\n"
+            flight_schedule_message += (
+                f"Flight from {flight.departure_airport} to {flight.destination_airport} at "
+                f"{flight.departure_time} UTC\n"
+            )
 
+        logger.debug("Sending new flights notification")
         self.send_notification(flight_schedule_message, NotificationLevel.INFO)
 
     def failed_reservation_retrieval(self, error: CheckInError, confirmation_number: str) -> None:
@@ -54,13 +63,15 @@ class NotificationHandler:
             f"with confirmation number {confirmation_number}. Reason: {error}.\n"
             f"Make sure the flight information is correct and try again.\n"
         )
+        logger.debug("Sending failed reservation retrieval notification...")
         self.send_notification(error_message, NotificationLevel.ERROR)
 
     def failed_login(self, error: LoginError) -> None:
         error_message = (
             f"Failed to log in to account with username {self.flight_retriever.username}. "
-            f"Status Code: {error}.\nMake sure the login information is correct and try again.\n"
+            f"{error}.\n"
         )
+        logger.debug("Sending failed login notification...")
         self.send_notification(error_message, NotificationLevel.ERROR)
 
     def successful_checkin(self, boarding_pass: Dict[str, Any], flight: Flight) -> None:
@@ -71,13 +82,19 @@ class NotificationHandler:
 
         for flight_info in boarding_pass["flights"]:
             for passenger in flight_info["passengers"]:
-                success_message += f"{passenger['name']} got {passenger['boardingGroup']}{passenger['boardingPosition']}!\n"
+                success_message += (
+                    f"{passenger['name']} got "
+                    f"{passenger['boardingGroup']}{passenger['boardingPosition']}!\n"
+                )
 
+        logger.debug("Sending successful check-in notification...")
         self.send_notification(success_message, NotificationLevel.INFO)
 
     def failed_checkin(self, error: CheckInError, flight: Flight) -> None:
         error_message = (
-            f"Failed to check in to flight {flight.confirmation_number} for {self._get_account_name()} "
-            f"Reason: {error}.\nCheck in at this url: {MANUAL_CHECKIN_URL}\n"
+            f"Failed to check in to flight {flight.confirmation_number} for "
+            f"{self._get_account_name()}. Reason: {error}.\nCheck in at this url: "
+            f"{MANUAL_CHECKIN_URL}\n"
         )
+        logger.debug("Sending failed check-in notification...")
         self.send_notification(error_message, NotificationLevel.ERROR)