+++ lib/webdriver.py
@@ -8,9 +8,11 @@ from typing import TYPE_CHECKING, Any, Dict
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
+from seleniumwire.request import Response
 from seleniumwire.undetected_chromedriver import Chrome, ChromeOptions
 
 from .general import LoginError
+from .log import get_logger
 
 if TYPE_CHECKING:  # pragma: no cover
     from .checkin_scheduler import CheckInScheduler
@@ -22,6 +24,11 @@ LOGIN_URL = BASE_URL + "/api/security/v4/security/token"
 TRIPS_URL = BASE_URL + "/api/mobile-misc/v1/mobile-misc/page/upcoming-trips"
 CHECKIN_URL = BASE_URL + "/check-in"
 RESERVATION_URL = BASE_URL + "/api/mobile-air-operations/v1/mobile-air-operations/page/check-in/"
+
+# Southwest's code when logging in with the incorrect information
+INVALID_CREDENTIALS_CODE = 400518024
+
+logger = get_logger(__name__)
 
 
 class WebDriver:
@@ -51,6 +58,7 @@ class WebDriver:
         headers from the request. Then, it updates the headers in the check-in scheduler.
         """
         driver = self._get_driver()
+        logger.debug("Filling out a check-in form to get valid headers")
 
         # Attempt a check in to retrieve the correct headers
         confirmation_element = WebDriverWait(driver, 30).until(
@@ -80,6 +88,7 @@ class WebDriver:
         information.
         """
         driver = self._get_driver()
+        logger.debug("Logging into Southwest account to get scheduled flights and valid headers")
 
         # Log in to retrieve the account's trips and needed headers for later requests
         WebDriverWait(driver, 30).until(
@@ -101,24 +110,30 @@ class WebDriver:
 
         response = driver.requests[0].response
         if response.status_code != 200:
-            raise LoginError(str(response.status_code))
+            error = self._handle_login_error(response)
+            raise error
 
-        # If this is the first time logging in, the account name needs to be set because that info is needed later
+        # If this is the first time logging in, the account name needs to be set
+        # because that info is needed later
         if flight_retriever.first_name is None:
+            logger.debug("First time logging in. Setting account name")
             response_body = json.loads(response.body)
             self._set_account_name(flight_retriever, response_body)
             print(
-                f"Successfully logged in to {flight_retriever.first_name} {flight_retriever.last_name}'s account\n"
-            )
+                f"Successfully logged in to {flight_retriever.first_name} "
+                f"{flight_retriever.last_name}'s account\n"
+            )  # Don't log as it contains sensitive information
 
-        # This page is also loaded when we log in, so we might as well grab it instead of requesting again later
+        # This page is also loaded when we log in, so we might as well grab it instead of
+        # requesting again later
         flights = json.loads(driver.requests[1].response.body)["upcomingTripsPage"]
 
         driver.quit()
 
-        return flights
+        return [flight for flight in flights if flight["tripType"] == "FLIGHT"]
 
     def _get_driver(self) -> Chrome:
+        logger.debug("Starting webdriver for current session")
         chrome_version = self.checkin_scheduler.flight_retriever.config.chrome_version
         driver = Chrome(
             options=self.options,
@@ -126,6 +141,8 @@ class WebDriver:
             version_main=chrome_version,
         )
         driver.scopes = [LOGIN_URL, TRIPS_URL, RESERVATION_URL]  # Filter out unneeded URLs
+
+        logger.debug("Loading Southwest Check-In page")
         driver.get(CHECKIN_URL)
         return driver
 
@@ -133,6 +150,7 @@ class WebDriver:
         # Retrieving the headers could fail if the form isn't given enough time to submit
         time.sleep(10)
 
+        logger.debug("Setting valid headers from previous request")
         request_headers = driver.requests[0].headers
         self.checkin_scheduler.headers = self._get_needed_headers(request_headers)
 
@@ -153,6 +171,18 @@ class WebDriver:
             options.add_argument("--headless=chrome")
 
         return options
+
+    @staticmethod
+    def _handle_login_error(response: Response) -> LoginError:
+        body = json.loads(response.body)
+        if body.get("code") == INVALID_CREDENTIALS_CODE:
+            logger.debug("Invalid credentials provided when attempting to log in")
+            reason = "Invalid credentials"
+        else:
+            logger.debug("Logging in failed for an unknown reason")
+            reason = "Unknown"
+
+        return LoginError(f"Reason: {reason}. Status code: {response.status_code}")
 
     @staticmethod
     def _get_needed_headers(request_headers: Dict[str, Any]) -> Dict[str, Any]: